requirements.txt
Django==2.2.16
Pillow==8.3.1
requests==2.26.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter==2.4.0
django-colorfield
webcolors
djoser
django-cors-headers